Tenging komin við mongodb
