+++ Website/Frammhaldskolaleitinn/client/src/SchoolPage.js
@@ -3,13 +3,13 @@ import React, { useEffect } from 'react';
 import { useParams, Link } from 'react-router-dom';
 
 const SchoolPage = ({ jsonData }) => {
-  const { id } = useParams();
+  const { nafn } = useParams();
 
   useEffect(() => {
     // Additional logic or side effects related to the school data can be added here
-  }, [id, jsonData]);
+  }, [nafn, jsonData]);
 
-  const school = jsonData ? jsonData.find((school) => school.ID.toString() === id) : null;
+  const school = jsonData ? jsonData.find((school) => school.nafn.toString() === nafn) : null;
 
   if (!school) {
     return <div>School not found</div>;
